process first round of review comments by @wcawijngaards

+++ testcode/testpkts.c
@@ -1525,7 +1525,7 @@ find_match(struct entry* entries, uint8_t* query_pkt, size_t len,
 			} else if(!p->match_ede_any &&
 				(uint16_t)info_code != p->ede_info_code) {
 				verbose(3, "bad EDE INFO-CODE. Expected: %d, "
-					"and got: %d\n", p->ede_info_code,
+					"and got: %d\n", (int)p->ede_info_code,
 					info_code);
 				continue;
 			}
